+++ runtime/datatypes/path.ts
@@ -1,5 +1,6 @@
 import * as Red from "../../red-types";
 import RedActions from "../actions";
+import {$$skip} from "./series";
 
 function insertOnly(
 	list:    Red.AnyType[],
@@ -104,12 +105,13 @@ export function $$append(
 // ...
 
 export function $$insert(
-	_ctx:  Red.Context,
+	ctx:   Red.Context,
 	path:  Red.RawAnyPath,
 	value: Red.AnyType,
 	_: RedActions.InsertOptions = {}
 ): typeof path {
 	const index = path.index - 1;
+	let offset = 0;
 	
 	if(_.dup !== undefined) {
 		const dups = [];
@@ -127,20 +129,79 @@ export function $$insert(
 		}
 		
 		path.path.splice(index, 0, ...dups);
-		path.index += dups.length;
+		offset += dups.length;
 	} else if(_.only !== undefined) {
 		insertOnly(path.path, value, index);
-		path.index++;
+		offset++;
 	} else {
 		if(value instanceof Red.RawBlock || value instanceof Red.RawHash || value instanceof Red.RawParen) {
-			path.index += insertAll(path.path, value.current().values, index, _.part);
+			offset += insertAll(path.path, value.current().values, index, _.part);
 		} else if(Red.isAnyPath(value)) {
-			path.index += insertAll(path.path, value.current().path, index, _.part);
+			offset += insertAll(path.path, value.current().path, index, _.part);
 		} else {
 			insertOnly(path.path, value, index);
-			path.index++;
+			offset++;
 		}
 	}
 
-	return path;
+	return $$skip(ctx, path, offset);
+}
+
+export function $$change(
+	ctx:   Red.Context,
+	path:  Red.RawAnyPath,
+	value: Red.AnyType,
+	_: RedActions.ChangeOptions = {}
+): typeof path {
+	const index = path.index - 1;
+	let offset = 0;
+	
+	if(_.dup !== undefined) {
+		const dups = [];
+		
+		if(_.only !== undefined) {
+			for(let i = 0; i < _.dup; i++) dups.push(value);
+		} else if(value instanceof Red.RawBlock || value instanceof Red.RawHash || value instanceof Red.RawParen) {
+			const values = value.current().values;
+			for(let i = 0; i < _.dup; i++) dups.push(...values);
+		} else if(Red.isAnyPath(value)) {
+			const values = value.current().path;
+			for(let i = 0; i < _.dup; i++) dups.push(...values);
+		} else {
+			for(let i = 0; i < _.dup; i++) dups.push(value);
+		}
+		
+		path.path.splice(index, dups.length, ...dups);
+		offset += dups.length;
+	} else if(_.only !== undefined) {
+		path.path[index] = value;
+		offset++;
+	} else {
+		if(value instanceof Red.RawBlock || value instanceof Red.RawHash || value instanceof Red.RawParen) {
+			const values = value.current().values;
+			
+			if(_.part === undefined) {
+				path.path.splice(index, values.length, ...values);
+				offset += values.length;
+			} else {
+				path.path.splice(index, _.part, ...values.slice(0, _.part));
+				offset += _.part;
+			}
+		} else if(Red.isAnyPath(value)) {
+			const values = value.current().path;
+			
+			if(_.part === undefined) {
+				path.path.splice(index, values.length, ...values);
+				offset += values.length;
+			} else {
+				path.path.splice(index, _.part, ...values.slice(0, _.part));
+				offset += _.part;
+			}
+		} else {
+			path.path[index] = value;
+			offset++;
+		}
+	}
+
+	return $$skip(ctx, path, offset);
 }