+++ runtime/datatypes/tag.ts
@@ -1,6 +1,7 @@
 import * as Red from "../../red-types";
 import RedActions from "../actions";
-import {append, insert} from "./string-ref";
+import {append, insert, change} from "./string-ref";
+import {$$skip} from "./series";
 
 /* Actions */
 export function $$form(
@@ -38,15 +39,20 @@ export function $$append(
 	return tag;
 }
 
-// ...
-
 export function $$insert(
 	ctx:   Red.Context,
 	tag:   Red.RawTag,
 	value: Red.AnyType,
 	_: RedActions.InsertOptions = {}
 ): Red.RawTag {
-	tag.index += insert(ctx, tag.tag, tag.index - 1, value, _);
-	
-	return tag;
+	return $$skip(ctx, tag, insert(ctx, tag.tag, tag.index - 1, value, _));
+}
+
+export function $$change(
+	ctx:   Red.Context,
+	tag:   Red.RawTag,
+	value: Red.AnyType,
+	_: RedActions.ChangeOptions = {}
+): Red.RawTag {
+	return $$skip(ctx, tag, change(ctx, tag.tag, tag.index - 1, value, _));
 }