+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Version 0.1.3 (in development)
 - Reimplemented `vector!`s.
 - Added `make` action for `vector!`s.
+- Add `change` action.
+- Fix `insert` action.
 
 
 # Version 0.1.2